# server/swagger-babel-node/config/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: My TODO APP
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /api/todos:
    # binds a127 app logic to a route
    x-swagger-router-controller: src/todos/routers/sub1/todo-router
    get:
      description: returns all Todos
      # used as the method name of the controller
      operationId: getAlltodos
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/TodoResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      summary: Create a new todo
      operationId: createTodo
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        201:
          description: Created
          schema:
            $ref: "#/definitions/TodoResponse"
      parameters:
        - name: "todo"
          in: body
          description: Todo to add to the system
          required: true
          schema:
            $ref: '#/definitions/NewTodo'
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  NewTodo:
    required:
      - completed
      - description
    properties:
      completed:
        type: boolean
      description:
        type: string
  
  TodoResponse:
    required:
      - id
      - completed
      - description
    properties:
      id:
        type: string
      completed:
        type: boolean
      description:
        type: string
  
  ErrorResponse:
    required:
      - message
      - code
    properties:
      message:
        type: string
      code:
        type: string